build: update query
