+++ pkg/logical/analyzer_test.go
@@ -29,6 +29,7 @@ import (
 	apiv1 "github.com/apache/skywalking-banyandb/api/fbs/v1"
 	apischema "github.com/apache/skywalking-banyandb/api/schema"
 	"github.com/apache/skywalking-banyandb/banyand/series"
+	"github.com/apache/skywalking-banyandb/pkg/fb"
 	"github.com/apache/skywalking-banyandb/pkg/logical"
 )
 
@@ -39,10 +40,10 @@ func TestAnalyzer_SimpleTimeScan(t *testing.T) {
 
 	sT, eT := time.Now().Add(-3*time.Hour), time.Now()
 
-	builder := NewCriteriaBuilder()
-	criteria := builder.Build(
-		AddLimit(0),
-		AddOffset(0),
+	builder := fb.NewCriteriaBuilder()
+	criteria := builder.BuildEntityCriteria(
+		fb.AddLimit(0),
+		fb.AddOffset(0),
 		builder.BuildMetaData("default", "trace"),
 		builder.BuildTimeStampNanoSeconds(sT, eT),
 	)
@@ -76,10 +77,10 @@ func TestAnalyzer_ComplexQuery(t *testing.T) {
 
 	sT, eT := time.Now().Add(-3*time.Hour), time.Now()
 
-	builder := NewCriteriaBuilder()
-	criteria := builder.Build(
-		AddLimit(5),
-		AddOffset(10),
+	builder := fb.NewCriteriaBuilder()
+	criteria := builder.BuildEntityCriteria(
+		fb.AddLimit(5),
+		fb.AddOffset(10),
 		builder.BuildMetaData("default", "trace"),
 		builder.BuildTimeStampNanoSeconds(sT, eT),
 		builder.BuildFields("service_id", "=", "my_app", "http.method", "=", "GET"),
@@ -121,10 +122,10 @@ func TestAnalyzer_TraceIDQuery(t *testing.T) {
 
 	ana := logical.DefaultAnalyzer()
 
-	builder := NewCriteriaBuilder()
-	criteria := builder.Build(
-		AddLimit(5),
-		AddOffset(10),
+	builder := fb.NewCriteriaBuilder()
+	criteria := builder.BuildEntityCriteria(
+		fb.AddLimit(5),
+		fb.AddOffset(10),
 		builder.BuildMetaData("default", "trace"),
 		builder.BuildFields("trace_id", "=", "123"),
 	)
@@ -151,10 +152,10 @@ func TestAnalyzer_Fields_FieldNotDefined(t *testing.T) {
 
 	ana := logical.DefaultAnalyzer()
 
-	builder := NewCriteriaBuilder()
-	criteria := builder.Build(
-		AddLimit(5),
-		AddOffset(10),
+	builder := fb.NewCriteriaBuilder()
+	criteria := builder.BuildEntityCriteria(
+		fb.AddLimit(5),
+		fb.AddOffset(10),
 		builder.BuildMetaData("default", "trace"),
 		builder.BuildTimeStampNanoSeconds(time.Now().Add(-3*time.Hour), time.Now()),
 		builder.BuildFields("duration", ">", 500),
@@ -179,10 +180,10 @@ func TestAnalyzer_OrderBy_FieldNotDefined(t *testing.T) {
 
 	ana := logical.DefaultAnalyzer()
 
-	builder := NewCriteriaBuilder()
-	criteria := builder.Build(
-		AddLimit(5),
-		AddOffset(10),
+	builder := fb.NewCriteriaBuilder()
+	criteria := builder.BuildEntityCriteria(
+		fb.AddLimit(5),
+		fb.AddOffset(10),
 		builder.BuildMetaData("default", "trace"),
 		builder.BuildTimeStampNanoSeconds(time.Now().Add(-3*time.Hour), time.Now()),
 		builder.BuildOrderBy("duration", apiv1.SortDESC),
@@ -206,10 +207,10 @@ func TestAnalyzer_Projection_FieldNotDefined(t *testing.T) {
 
 	ana := logical.DefaultAnalyzer()
 
-	builder := NewCriteriaBuilder()
-	criteria := builder.Build(
-		AddLimit(5),
-		AddOffset(10),
+	builder := fb.NewCriteriaBuilder()
+	criteria := builder.BuildEntityCriteria(
+		fb.AddLimit(5),
+		fb.AddOffset(10),
 		builder.BuildMetaData("default", "trace"),
 		builder.BuildTimeStampNanoSeconds(time.Now().Add(-3*time.Hour), time.Now()),
 		builder.BuildProjection("duration", "service_id"),
@@ -232,10 +233,10 @@ func TestAnalyzer_Fields_IndexNotDefined(t *testing.T) {
 
 	ana := logical.DefaultAnalyzer()
 
-	builder := NewCriteriaBuilder()
-	criteria := builder.Build(
-		AddLimit(5),
-		AddOffset(10),
+	builder := fb.NewCriteriaBuilder()
+	criteria := builder.BuildEntityCriteria(
+		fb.AddLimit(5),
+		fb.AddOffset(10),
 		builder.BuildMetaData("default", "trace"),
 		builder.BuildFields("service_name", "=", "app"),
 		builder.BuildTimeStampNanoSeconds(time.Now().Add(-3*time.Hour), time.Now()),
